requirements.txt
transformers==4.36.0
accelerate==0.24.0
librosa==0.10.0
torch==2.0.1
torchaudio==2.0.2
huggingface-hub==0.19.0
fastapi==0.104.0
uvicorn==0.24.0
pyngrok==6.0.0
nest-asyncio==1.5.8
python-dotenv==1.0.0
pyaudio==0.2.13
requests==2.31.0
keyboard==0.13.5
